Finish core + add commander + fix execa conflition with cjs module
Execa build with ejs module, so if we build our project with commonjs
module system, the output build artifact would not be able to import
execa ejs module, then broken and throw error of EJS import.

So our solution is downgrading the execa instance to 5x version, that
still uses commonjs. Because the alternation solution of converting
our build artifact from cjs to ejs is so intricated.